Query httpgd plot IDs

Query httpgd graphics device static plot IDs. Available plot IDs starting from index will be returned. limit specifies the number of plots. This function will only work after starting a device with hgd().

Usage

hgd_id(index = 0, limit = 1, which = dev.cur(), state = FALSE)

Arguments

index

Plot index. If this is set to 0, the last page will be selected.

limit

Limit the number of returned IDs. If this is set to a value > 1 the returned type is a list if IDs.

which

Which device (ID).

state

Include the current device state in the returned result (see also: hgd_state()).

Value

TODO

Examples

if (FALSE) {

hgd()
plot.new()
text(.5, .5, "#1")
plot.new()
text(.5, .5, "#2")
plot.new()
text(.5, .5, "#3")
third <- hgd_id()
second <- hgd_id(2)
all <- hgd_id(1, limit = Inf)
hgd_remove(1)
hgd_svg(second)

dev.off()
}