Inline SVG rendering.

Convenience function for quick inline SVG rendering. This is similar to hgd_svg() but the plotting code is specified inline and an offline httpgd graphics device is managed (created and closed) automatically. Starting a device with hgd() is therefore not necessary.

Usage

hgd_inline(
  code,
  page = 0,
  page_width = -1,
  page_height = -1,
  zoom = 1,
  renderer = "svg",
  file = NA,
  ...
)

Arguments

code

Plotting code. See examples for more information.

page

Plot page to render. If this is set to 0, the last page will be selected. Can be set to a numeric plot index or plot ID (see hgd_id()).

page_width

Width of the plot. If this is set to -1, the last width will be selected.

page_height

Height of the plot. If this is set to -1, the last height will be selected.

zoom

Zoom level. (For example: 2 corresponds to 200%, 0.5 would be 50%.)

renderer

Renderer.

file

Filepath to save SVG. (No file will be created if this is NA)

...

Additional parameters passed to hgd(webserver=FALSE, ...)

Value

Rendered SVG string.

Examples

hgd_inline({
  hist(rnorm(100))
})
#> [1] "<svg xmlns=\"http://www.w3.org/2000/svg\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" class=\"httpgd\" width=\"720.00\" height=\"576.00\" viewBox=\"0 0 720.00 576.00\">\n<defs>\n  <style type='text/css'><![CDATA[\n    .httpgd line, .httpgd polyline, .httpgd polygon, .httpgd path, .httpgd rect, .httpgd circle {\n      fill: none;\n      stroke: #000000;\n      stroke-linecap: round;\n      stroke-linejoin: round;\n      stroke-miterlimit: 10.00;\n    }\n  ]]></style>\n<clipPath id=\"c0\"><rect x=\"0.00\" y=\"0.00\" width=\"720.00\" height=\"576.00\"/></clipPath>\n<clipPath id=\"c1\"><rect x=\"59.04\" y=\"59.04\" width=\"630.72\" height=\"443.52\"/></clipPath>\n</defs>\n<rect width=\"100%\" height=\"100%\" style=\"stroke: none;fill: #FFFFFF;\"/>\n<g clip-path=\"url(#c0)\">\n<g><text x=\"374.40\" y=\"34.77\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 14.40px;font-weight: bold;\" textLength=\"203.13px\" lengthAdjust=\"spacingAndGlyphs\">Histogram of rnorm(100)</text></g>\n<g><text x=\"374.40\" y=\"557.28\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"68.37px\" lengthAdjust=\"spacingAndGlyphs\">rnorm(100)</text></g>\n<g><text transform=\"translate(12.96,280.80) rotate(-90.00)\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"62.02px\" lengthAdjust=\"spacingAndGlyphs\">Frequency</text></g>\n<line x1=\"131.07\" y1=\"502.56\" x2=\"617.73\" y2=\"502.56\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"131.07\" y1=\"502.56\" x2=\"131.07\" y2=\"509.76\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"228.40\" y1=\"502.56\" x2=\"228.40\" y2=\"509.76\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"325.73\" y1=\"502.56\" x2=\"325.73\" y2=\"509.76\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"423.07\" y1=\"502.56\" x2=\"423.07\" y2=\"509.76\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"520.40\" y1=\"502.56\" x2=\"520.40\" y2=\"509.76\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"617.73\" y1=\"502.56\" x2=\"617.73\" y2=\"509.76\" style=\"stroke-width: 0.75;\"/>\n<g><text x=\"131.07\" y=\"528.48\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"11.97px\" lengthAdjust=\"spacingAndGlyphs\">-3</text></g>\n<g><text x=\"228.40\" y=\"528.48\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"11.97px\" lengthAdjust=\"spacingAndGlyphs\">-2</text></g>\n<g><text x=\"325.73\" y=\"528.48\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"11.97px\" lengthAdjust=\"spacingAndGlyphs\">-1</text></g>\n<g><text x=\"423.07\" y=\"528.48\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"7.64px\" lengthAdjust=\"spacingAndGlyphs\">0</text></g>\n<g><text x=\"520.40\" y=\"528.48\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"7.64px\" lengthAdjust=\"spacingAndGlyphs\">1</text></g>\n<g><text x=\"617.73\" y=\"528.48\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"7.64px\" lengthAdjust=\"spacingAndGlyphs\">2</text></g>\n<line x1=\"59.04\" y1=\"486.13\" x2=\"59.04\" y2=\"129.03\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"59.04\" y1=\"486.13\" x2=\"51.84\" y2=\"486.13\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"59.04\" y1=\"396.86\" x2=\"51.84\" y2=\"396.86\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"59.04\" y1=\"307.58\" x2=\"51.84\" y2=\"307.58\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"59.04\" y1=\"218.31\" x2=\"51.84\" y2=\"218.31\" style=\"stroke-width: 0.75;\"/>\n<line x1=\"59.04\" y1=\"129.03\" x2=\"51.84\" y2=\"129.03\" style=\"stroke-width: 0.75;\"/>\n<g><text transform=\"translate(41.76,486.13) rotate(-90.00)\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"7.64px\" lengthAdjust=\"spacingAndGlyphs\">0</text></g>\n<g><text transform=\"translate(41.76,396.86) rotate(-90.00)\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"7.64px\" lengthAdjust=\"spacingAndGlyphs\">5</text></g>\n<g><text transform=\"translate(41.76,307.58) rotate(-90.00)\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"15.28px\" lengthAdjust=\"spacingAndGlyphs\">10</text></g>\n<g><text transform=\"translate(41.76,218.31) rotate(-90.00)\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"15.28px\" lengthAdjust=\"spacingAndGlyphs\">15</text></g>\n<g><text transform=\"translate(41.76,129.03) rotate(-90.00)\" text-anchor=\"middle\" style=\"font-family: Liberation Sans;font-size: 12.00px;\" textLength=\"15.28px\" lengthAdjust=\"spacingAndGlyphs\">20</text></g>\n</g><g clip-path=\"url(#c1)\">\n<rect x=\"82.40\" y=\"468.28\" width=\"48.67\" height=\"17.86\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"131.07\" y=\"486.13\" width=\"48.67\" height=\"0.00\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"179.73\" y=\"450.42\" width=\"48.67\" height=\"35.71\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"228.40\" y=\"396.86\" width=\"48.67\" height=\"89.28\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"277.07\" y=\"307.58\" width=\"48.67\" height=\"178.55\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"325.73\" y=\"271.87\" width=\"48.67\" height=\"214.26\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"374.40\" y=\"75.47\" width=\"48.67\" height=\"410.67\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"423.07\" y=\"218.31\" width=\"48.67\" height=\"267.83\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"471.73\" y=\"236.16\" width=\"48.67\" height=\"249.97\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"520.40\" y=\"307.58\" width=\"48.67\" height=\"178.55\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"569.07\" y=\"361.15\" width=\"48.67\" height=\"124.99\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n<rect x=\"617.73\" y=\"468.28\" width=\"48.67\" height=\"17.86\" style=\"stroke-width: 0.75;fill: #D3D3D3;\"/>\n</g>\n</svg>"

s <- hgd_inline({
  plot.new()
  lines(c(0.5, 1, 0.5), c(0.5, 1, 1))
})
cat(s)
#> <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" class="httpgd" width="720.00" height="576.00" viewBox="0 0 720.00 576.00">
#> <defs>
#>   <style type='text/css'><![CDATA[
#>     .httpgd line, .httpgd polyline, .httpgd polygon, .httpgd path, .httpgd rect, .httpgd circle {
#>       fill: none;
#>       stroke: #000000;
#>       stroke-linecap: round;
#>       stroke-linejoin: round;
#>       stroke-miterlimit: 10.00;
#>     }
#>   ]]></style>
#> <clipPath id="c0"><rect x="0.00" y="0.00" width="720.00" height="576.00"/></clipPath>
#> <clipPath id="c1"><rect x="59.04" y="59.04" width="630.72" height="443.52"/></clipPath>
#> </defs>
#> <rect width="100%" height="100%" style="stroke: none;fill: #FFFFFF;"/>
#> <g clip-path="url(#c0)">
#> </g><g clip-path="url(#c1)">
#> <polyline points="374.40,280.80 666.40,75.47 374.40,75.47" style="stroke-width: 0.75;"/>
#> </g>
#> </svg>